[~][=] split ffio.py
